Refactored project structure for the purposes of this engine-focused fork
