+++ EngineTests/src/main/java/games/temporalstudio/enginetests/Launcher.java
@@ -1,0 +1,7 @@
+package games.temporalstudio.enginetests;
+
+public class Launcher {
+	public static void main(String[] args) {
+		new TestGame().run(new String[0]);
+	}
+}
