+++ EngineTests/src/main/resources/assets/maps/tiles.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.8" tiledversion="1.8.2" name="tiles" tilewidth="64" tileheight="64" tilecount="150" columns="15">
+ <image source="../textures/past.png" width="960" height="640"/>
+</tileset>
