+++ EngineTests/src/main/resources/assets/textures/bat.json
@@ -1,0 +1,11 @@
+{
+	"fly": {
+		"positions": [
+			[0, 0],
+			[1, 0],
+			[2, 0],
+			[3, 0]
+		],
+		"scale": [1, 1]
+	}
+}
