+++ EngineTests/src/main/resources/assets/textures/jeanne.json
@@ -1,0 +1,74 @@
+{
+	"face_walk": {
+		"positions": [
+			[0, 0],
+			[1, 0],
+			[2, 0],
+			[3, 0]
+		],
+		"scale": [1, 2]
+	},
+	"face_jump": {
+		"positions": [
+			[4, 0],
+			[5, 0],
+			[6, 0],
+			[7, 0]
+		],
+		"scale": [1, 2]
+	},
+	"right_walk": {
+		"positions": [
+			[0, 2],
+			[1, 2],
+			[2, 2],
+			[3, 2]
+		],
+		"scale": [1, 2]
+	},
+	"right_jump": {
+		"positions": [
+			[4, 2],
+			[5, 2],
+			[6, 2],
+			[7, 2]
+		],
+		"scale": [1, 2]
+	},
+	"left_walk": {
+		"positions": [
+			[0, 4],
+			[1, 4],
+			[2, 4],
+			[3, 4]
+		],
+		"scale": [1, 2]
+	},
+	"left_jump": {
+		"positions": [
+			[4, 4],
+			[5, 4],
+			[6, 4],
+			[7, 4]
+		],
+		"scale": [1, 2]
+	},
+	"back_walk": {
+		"positions": [
+			[0, 6],
+			[1, 6],
+			[2, 6],
+			[3, 6]
+		],
+		"scale": [1, 2]
+	},
+	"back_jump": {
+		"positions": [
+			[4, 6],
+			[5, 6],
+			[6, 6],
+			[7, 6]
+		],
+		"scale": [1, 2]
+	}
+}
