+++ EngineTests/src/main/resources/assets/textures/meduse.json
@@ -1,0 +1,38 @@
+{
+	"back_walk": {
+		"positions": [
+			[0, 0],
+			[1, 0],
+			[2, 0],
+			[3, 0]
+		],
+		"scale": [1, 1]
+	},
+	"face_walk": {
+		"positions": [
+			[0, 1],
+			[1, 1],
+			[2, 1],
+			[3, 1]
+		],
+		"scale": [1, 1]
+	},
+	"left_walk": {
+		"positions": [
+			[0, 2],
+			[1, 2],
+			[2, 2],
+			[3, 2]
+		],
+		"scale": [1, 1]
+	},
+	"right_walk": {
+		"positions": [
+			[0, 3],
+			[1, 3],
+			[2, 3],
+			[3, 3]
+		],
+		"scale": [1, 1]
+	}
+}
