+++ EngineTests/src/main/resources/assets/textures/merlin.json
@@ -1,0 +1,34 @@
+{
+	"face_walk": {
+		"positions" : [[0,0],[1,0],[2,0],[3,0]],
+		"scale" : [1, 2]
+	},
+	"face_jump": {
+		"positions" : [[0,2],[1,2],[2,2],[3,2]],
+		"scale" : [1, 2]
+	},
+	"back_walk": {
+		"positions" : [[0,4],[1,4],[2,4],[3,4]],
+		"scale" : [1, 2]
+	},
+	"back_jump": {
+		"positions" : [[0,6],[1,6],[2,6],[3,6]],
+		"scale" : [1, 2]
+	},
+	"right_walk": {
+		"positions" : [[0,8],[1,8],[2,8],[3,8]],
+		"scale" : [1, 2]
+	},
+	"right_jump": {
+		"positions" : [[0,10],[1,10],[2,10],[3,10]],
+		"scale" : [1, 2]
+	},
+	"left_walk": {
+		"positions" : [[0,12],[1,12],[2,12],[3,12]],
+		"scale" : [1, 2]
+	},
+	"left_jump": {
+		"positions" : [[0,14],[1,14],[2,14],[3,14]],
+		"scale" : [1, 2]
+	}
+}
