+++ EngineTests/src/main/resources/assets/textures/past.json
@@ -1,0 +1,207 @@
+{
+	"full_soil": {
+		"position": [0, 0],
+		"scale": [1, 1],
+		"tiledID": 0
+	},
+	"soil_and_right_river": {
+		"position": [1, 0],
+		"scale": [1, 1],
+		"tiledID": 1
+	},
+	"soil_and_left_river": {
+		"position": [2, 0],
+		"scale": [1, 1],
+		"tiledID": 2
+	},
+	"soil_and_bottom_river": {
+		"position": [3, 0],
+		"scale": [1, 1],
+		"tiledID": 3
+	},
+	"soil_and_top_river": {
+		"position": [4, 0],
+		"scale": [1, 1],
+		"tiledID": 4
+	},
+	"little_angle_top_right": {
+		"position": [5, 0],
+		"scale": [1, 1],
+		"tiledID": 5
+	},
+	"little_angle_top_left": {
+		"position": [6, 0],
+		"scale": [1, 1],
+		"tiledID": 6
+	},
+	"little_angle_bottom_right": {
+		"position": [7, 0],
+		"scale": [1, 1],
+		"tiledID": 7
+	},
+	"little_angle_bottom_left": {
+		"position": [8, 0],
+		"scale": [1, 1],
+		"tiledID": 8
+	},
+	"big_angle_top_right": {
+		"position": [9, 0],
+		"scale": [1, 1],
+		"tiledID": 9
+	},
+	"big_angle_top_left": {
+		"position": [10, 0],
+		"scale": [1, 1],
+		"tiledID": 10
+	},
+	"big_angle_bottom_left": {
+		"position": [11, 0],
+		"scale": [1, 1],
+		"tiledID": 11
+	},
+	"big_angle_bottom_right": {
+		"position": [12, 0],
+		"scale": [1, 1],
+		"tiledID": 12
+	},
+	"full_water": {
+		"position": [13, 0],
+		"scale": [1, 1],
+		"tiledID": 13
+	},
+	"little_seed_emplacement": {
+		"position": [14, 0],
+		"scale": [1,1]
+	},
+
+
+	"full_wall1": {
+		"position": [0, 1],
+		"scale": [1, 3],
+		"tiledID": 45
+	},
+	"full_wall2": {
+		"position": [1, 1],
+		"scale": [1, 3],
+		"tiledID": 46
+	},
+	"end_right": {
+		"position": [2, 1],
+		"scale": [1, 3],
+		"tiledID": 47
+	},
+	"end_left": {
+		"position": [3, 1],
+		"scale": [1, 3],
+		"tiledID": 48
+	},
+	"back_wall": {
+		"position": [4, 1],
+		"scale": [1, 3]
+	},
+	"left_wall": {
+		"position": [5, 1],
+		"scale": [1, 3],
+		"tiledID": 50
+	},
+	"right_wall": {
+		"position": [6, 1],
+		"scale": [1, 3],
+		"tiledID": 51
+	},
+	"left_wall_0": {
+		"position": [5, 1],
+		"scale": [1, 1],
+		"tiledID": 20
+	},
+	"right_wall_0": {
+		"position": [6, 1],
+		"scale": [1, 1],
+		"tiledID": 21
+	},
+	"bottom_left_coin_wall": {
+		"position": [7, 1],
+		"scale": [1, 3],
+		"tiledID": 52
+	},
+	"bottom_right_coin_wall": {
+		"position": [8, 1],
+		"scale": [1, 3],
+		"tiledID": 53
+	},
+	"top_left_coin_wall": {
+		"position": [9, 1],
+		"scale": [1, 4]
+	},
+	"top_right_coin_wall": {
+		"position": [10, 1],
+		"scale": [1, 4]
+	},
+	"top_left_coin_wall_0-3": {
+		"position": [9, 1],
+		"scale": [1, 3],
+		"tiledID": 54
+	},
+	"top_right_coin_wall_0-3": {
+		"position": [10, 1],
+		"scale": [1, 3],
+		"tiledID": 55
+	},
+	"left_door": {
+		"position": [11, 1],
+		"scale": [1, 4]
+	},
+	"right_door": {
+		"position": [12, 1],
+		"scale": [1, 4]
+	},
+	"face_door": {
+		"position": [13, 1],
+		"scale": [2, 3]
+	},
+
+	"complete_key": {
+		"position": [0, 4],
+		"scale": [1, 1]
+	},
+	"key_fragment1": {
+		"position": [1, 4],
+		"scale": [1, 1]
+	},
+	"key_fragment2": {
+		"position": [2, 4],
+		"scale": [1, 1]
+	},
+	"seed": {
+		"position": [3, 4],
+		"scale": [1, 1]
+	},
+	"kiwi": {
+		"position": [4, 4],
+		"scale": [1, 1]
+	},
+	"small_root_bridge": {
+		"position": [0, 5],
+		"scale": [1, 3]
+	},
+	"big_root_bridge": {
+		"position": [1, 5],
+		"scale": [1, 4]
+	},
+	"bottle": {
+		"position": [2, 5],
+		"scale": [1, 4]
+	},
+	"tree": {
+		"position": [5, 4],
+		"scale": [3, 6]
+	},
+	"chest": {
+		"position": [8, 5],
+		"scale": [2, 2]
+	},
+	"big_seed_emplacement": {
+		"position": [10, 5],
+		"scale": [2, 2]
+	}
+}
