+++ EngineTests/src/main/resources/enginetests.properties
@@ -1,0 +1,2 @@
+name=enginetests
+version=1.0.0
