+++ EngineTests/src/main/resources/i18n/en/general.properties
@@ -1,0 +1,1 @@
+game.title=Engine Tests - v%s
