+++ EngineTests/src/main/resources/temporalengine.properties
@@ -1,0 +1,2 @@
+name=temporalengine
+version=1.0.0
