+++ build.gradle
@@ -89,13 +89,30 @@ project(":TimeCapsule"){
 		mainClass = "games.temporalstudio.timecapsule.Launcher"
 	}
 
-	task runEngine(type: JavaExec){
-		group = "application"
-		description = "Runs engine test part this project as a JVM application."
+	distributions{
+		main{
+			distributionBaseName = appName
+		}
+	}
+}
 
-		mainClass = application.mainClass
-    	classpath = sourceSets.main.runtimeClasspath
-		args = ["engine"]
+project(":EngineTests"){
+	apply plugin: "application"
+
+	def appName = "Engine Tests"
+	version = "1.0.0"
+
+	base{
+		archivesName = appName
+	}
+
+	dependencies{
+		implementation project(':TemporalEngine')
+	}
+
+	application{
+		applicationName = appName
+		mainClass = "games.temporalstudio.enginetests.Launcher"
 	}
 
 	distributions{
