+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'TemporalEngine'
 
-include("TemporalEngine", "TimeCapsule")
+include("TemporalEngine", "TimeCapsule", "EngineTests")
